sc5xx: move/rename asm headers
Move the sc5xx-related asm headers from:
	arch/arm/include/asm/mach-adi
to:
	arch/arm/include/asm/arch-adi/sc5xx
	arch/arm/include/asm/arch-adi/sc5xx-64
