+++ arch/arm/include/asm/arch-adi/sc5xx/adi_dma.h
