+++ arch/arm/include/asm/arch-adi/sc5xx/adi_mdma.h
@@ -11,7 +11,7 @@
 #ifndef ADI_MDMA
 #define ADI_MDMA
 
-#include <asm/mach-adi/common/adi_dma.h>
+#include <asm/arch-adi/sc5xx/adi_dma.h>
 
 #ifdef CONFIG_SC59X
     #include <asm/armv7.h>
