+++ arch/arm/include/asm/arch-adi/sc5xx/ospi.h
