+++ arch/arm/include/asm/arch-adi/sc5xx/sc5xx.h
