+++ board/adi/common/init/dmc/dmcinit.c
@@ -10,7 +10,7 @@
 
 #include <config.h>
 #include <common.h>
-#include <asm/mach-adi/common/sc5xx.h>
+#include <asm/arch-adi/sc5xx/sc5xx.h>
 #include "dmcinit.h"
 #include <stdbool.h>
