+++ board/adi/common/sc59x/sc59x-shared.c
@@ -17,7 +17,7 @@
 #include <asm/io.h>
 #include <asm/gpio.h>
 #include <asm/mach-types.h>
-#include <asm/mach-adi/common/sc5xx.h>
+#include <asm/arch-adi/sc5xx/sc5xx.h>
 #include <linux/delay.h>
 #include <watchdog.h>
 #include <asm/armv8/mmu.h>
