+++ board/adi/common/soc.c
@@ -9,7 +9,7 @@
  */
 
 #include <common.h>
-#include <asm/mach-adi/common/sc5xx.h>
+#include <asm/arch-adi/sc5xx/sc5xx.h>
 #include <asm/io.h>
 #include <asm/system.h>
 #include <cpu_func.h>
