+++ board/adi/sc584-ezkit/sc584-ezkit.c
@@ -14,7 +14,7 @@
 #include <asm/io.h>
 #include <asm/gpio.h>
 #include <asm/mach-types.h>
-#include <asm/mach-adi/common/sc5xx.h>
+#include <asm/arch-adi/sc5xx/sc5xx.h>
 #include <linux/delay.h>
 #include <watchdog.h>
