+++ board/adi/sc598-som-ezkit/sc598-som-ezkit.c
@@ -13,7 +13,7 @@
 #include <phy.h>
 #include <asm/io.h>
 #include <asm/mach-types.h>
-#include <asm/mach-adi/common/sc5xx.h>
+#include <asm/arch-adi/sc5xx/sc5xx.h>
 #include <linux/delay.h>
 #include <watchdog.h>
 #include <asm/armv8/mmu.h>
